+++ src/components/Feed/NotesFeed/components/NotesFeedTabs.tsx
@@ -28,7 +28,7 @@ const NotesFeedTabs: React.FC<Props> = ({ activeTab, setActiveTab }) => {
           },
         }}
       >
-        <Tab label="Discover" value="discover" /> {/* 🆕 */}
+        <Tab label="Network" value="discover" /> {/* 🆕 */}
         <Tab label="Following" value="following" />
         <Tab label="Reacted by Contacts" value="reacted" />
       </Tabs>